requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyYAML==6.0.1
pytest==7.4.3
